requirements.txt
dependency-injector==4.43.0
faiss-cpu==1.9.0
fastapi==0.115.4
openai==0.28
python-dotenv==1.0.1
pydantic==2.9.2
numpy==2.1.3
scikit-learn==1.5.2
sentence-transformers==3.3.0
uvicorn==0.32.0

